results/rebuttal_new_baselines/run-20231112-040700_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-finance-donnemartin-system-design-oop-bookclub_0_20-runs_log/finance/12/SelfRefine/app.py
from flask import Flask, request, jsonify
from dataclasses import dataclass

app = Flask(__name__)

# Mock database
DATABASE = {}

@dataclass
class User:
	username: str
	password: str

@dataclass
class Transaction:
	user: str
	category: str
	amount: float

@app.route('/create_user', methods=['POST'])
def create_user():
	data = request.get_json()
	user = User(**data)
	if user.username in DATABASE:
		return jsonify({'message': 'User already exists'}), 400
	DATABASE[user.username] = user
	return jsonify({'message': 'User created'}), 201

@app.route('/add_transaction', methods=['POST'])
def add_transaction():
	data = request.get_json()
	transaction = Transaction(**data)
	if transaction.user not in DATABASE:
		return jsonify({'message': 'User not found'}), 404
	if 'transactions' not in DATABASE[transaction.user].__dict__:
		DATABASE[transaction.user].__dict__['transactions'] = []
	DATABASE[transaction.user].__dict__['transactions'].append(transaction)
	return jsonify({'message': 'Transaction added'}), 201

if __name__ == '__main__':
	app.run(debug=True)
